move deploy stuff to src/echodataflow/deployment
